refactor(frontend): move results into a new `BlockSearchResultItem` component
